Move customer entity id to the checkout store
