+++ web/app/containers/checkout-shipping/actions.js
@@ -4,7 +4,7 @@ import CheckoutShipping from './container'
 import checkoutShippingParser from './parsers/checkout-shipping'
 import shippingMethodParser from './parsers/shipping-method'
 import {addNotification, fetchPage, removeAllNotifications, removeNotification} from '../app/actions'
-import {getCustomerEntityID} from './selectors'
+import {getCustomerEntityID} from '../../store/checkout/selectors'
 import {getIsLoggedIn} from '../app/selectors'
 import {getShippingFormValues} from '../../store/form/selectors'
 import {receiveShippingMethodInitialValues} from '../../store/checkout/actions'
