+++ web/app/store/checkout/actions.js
@@ -1,5 +1,5 @@
 import {createAction} from '../../utils/utils'
-import {extractMagentoShippingStepData} from '../../utils/magento-utils'
+import {extractMagentoShippingStepData, getCheckoutConfigObject} from '../../utils/magento-utils'
 
 import parseLocations from './locations/parser'
 import parseShippingInitialValues from './shipping/parser'
@@ -10,10 +10,16 @@ export const receiveShippingMethodInitialValues = createAction('Receive Shipping
 
 export const processCheckoutData = ({payload: {$response}}) => {
     return (dispatch) => {
+        const configObject = getCheckoutConfigObject($response)
         const magentoFieldData = extractMagentoShippingStepData($response).getIn(['children', 'shipping-address-fieldset', 'children'])
         const initialValues = parseShippingInitialValues(magentoFieldData)
         const locationsData = parseLocations(magentoFieldData)
 
-        dispatch(receiveCheckoutData({...locationsData, shipping: {initialValues}}))
+        dispatch(receiveCheckoutData({
+            ...locationsData,
+            shipping: {initialValues},
+            // entity_id is used for calls to the ID
+            customerEntityID: configObject ? configObject.quoteData.entity_id : ''
+        }))
     }
 }
