+++ web/app/store/checkout/selectors.js
@@ -1,0 +1,4 @@
+import {getCheckout} from '../../store/selectors'
+import {createGetSelector} from '../../utils/selector-utils'
+
+export const getCustomerEntityID = createGetSelector(getCheckout, 'customerEntityID')
